feat: cleanup all feature schemas, rename  ->
