+++ src/mcp/tools/featureTools.ts
@@ -7,17 +7,6 @@ import {
     deleteFeature,
     getFeatureAuditLogHistory,
 } from '../../api/features'
-import {
-    fetchVariations,
-    createVariation,
-    updateVariation,
-} from '../../api/variations'
-import {
-    enableTargeting,
-    disableTargeting,
-    fetchTargetingForFeature,
-    updateFeatureConfigForEnvironment,
-} from '../../api/targeting'
 
 import {
     ListFeaturesArgsSchema,
@@ -25,12 +14,6 @@ import {
     UpdateFeatureArgsSchema,
     UpdateFeatureStatusArgsSchema,
     DeleteFeatureArgsSchema,
-    SetFeatureTargetingArgsSchema,
-    ListVariationsArgsSchema,
-    CreateVariationArgsSchema,
-    UpdateVariationArgsSchema,
-    ListFeatureTargetingArgsSchema,
-    UpdateFeatureTargetingArgsSchema,
     GetFeatureAuditLogHistoryArgsSchema,
 } from '../types'
 import { IDevCycleApiClient } from '../api/interface'
@@ -178,219 +161,6 @@ export async function deleteFeatureHandler(
             }
         },
         dashboardLinks.feature.list,
-    )
-}
-
-export async function fetchFeatureVariationsHandler(
-    args: z.infer<typeof ListVariationsArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    return await apiClient.executeWithDashboardLink(
-        'fetchFeatureVariations',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please select a project using the select_project tool first.',
-                )
-            }
-            return await handleZodiosValidationErrors(
-                () => fetchVariations(authToken, projectKey, args.feature_key),
-                'fetchVariations',
-            )
-        },
-        (orgId: string, projectKey: string | undefined) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                args.feature_key,
-                'overview',
-            ),
-    )
-}
-
-export async function createFeatureVariationHandler(
-    args: z.infer<typeof CreateVariationArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    const { feature_key, ...variationData } = args
-
-    return await apiClient.executeWithDashboardLink(
-        'createFeatureVariation',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please select a project using the select_project tool first.',
-                )
-            }
-            return await handleZodiosValidationErrors(
-                () =>
-                    createVariation(
-                        authToken,
-                        projectKey,
-                        feature_key,
-                        variationData,
-                    ),
-                'createVariation',
-            )
-        },
-        (orgId: string, projectKey: string | undefined, result: any) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                result.key,
-                'manage-feature',
-            ),
-    )
-}
-
-export async function updateFeatureVariationHandler(
-    args: z.infer<typeof UpdateVariationArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    const { feature_key, variation_key, ...variationData } = args
-
-    return await apiClient.executeWithDashboardLink(
-        'updateFeatureVariation',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please select a project using the select_project tool first.',
-                )
-            }
-            return await handleZodiosValidationErrors(
-                () =>
-                    updateVariation(
-                        authToken,
-                        projectKey,
-                        feature_key,
-                        variation_key,
-                        variationData,
-                    ),
-                'updateVariation',
-            )
-        },
-        (orgId: string, projectKey: string | undefined, result: any) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                result.key,
-                'manage-feature',
-            ),
-    )
-}
-
-export async function setFeatureTargetingHandler(
-    args: z.infer<typeof SetFeatureTargetingArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    const operation = args.enabled ? 'enableTargeting' : 'disableTargeting'
-    const apiFunction = args.enabled ? enableTargeting : disableTargeting
-
-    return await apiClient.executeWithDashboardLink(
-        'set_feature_targeting',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please selecting a project first.',
-                )
-            }
-            await handleZodiosValidationErrors(
-                () =>
-                    apiFunction(
-                        authToken,
-                        projectKey,
-                        args.feature_key,
-                        args.environment_key,
-                    ),
-                operation,
-            )
-            const action = args.enabled ? 'enabled' : 'disabled'
-            return {
-                message: `Targeting ${action} for feature '${args.feature_key}' in environment '${args.environment_key}'`,
-            }
-        },
-        (orgId: string, projectKey: string | undefined) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                args.feature_key,
-                'manage-feature',
-            ),
-    )
-}
-
-export async function listFeatureTargetingHandler(
-    args: z.infer<typeof ListFeatureTargetingArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    return await apiClient.executeWithDashboardLink(
-        'listFeatureTargeting',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please select a project using the select_project tool first.',
-                )
-            }
-            return await handleZodiosValidationErrors(
-                () =>
-                    fetchTargetingForFeature(
-                        authToken,
-                        projectKey,
-                        args.feature_key,
-                        args.environment_key,
-                    ),
-                'fetchTargetingForFeature',
-            )
-        },
-        (orgId: string, projectKey: string | undefined) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                args.feature_key,
-                'manage-feature',
-            ),
-    )
-}
-
-export async function updateFeatureTargetingHandler(
-    args: z.infer<typeof UpdateFeatureTargetingArgsSchema>,
-    apiClient: IDevCycleApiClient,
-) {
-    const { feature_key, environment_key, ...configData } = args
-
-    return await apiClient.executeWithDashboardLink(
-        'update_feature_targeting',
-        args,
-        async (authToken: string, projectKey: string | undefined) => {
-            if (!projectKey) {
-                throw new Error(
-                    'Project key is required for this operation. Please select a project using the select_project tool first.',
-                )
-            }
-            return await handleZodiosValidationErrors(
-                () =>
-                    updateFeatureConfigForEnvironment(
-                        authToken,
-                        projectKey,
-                        feature_key,
-                        environment_key,
-                        configData,
-                    ),
-                'updateFeatureConfigForEnvironment',
-            )
-        },
-        (orgId: string, projectKey: string | undefined) =>
-            dashboardLinks.feature.dashboard(
-                orgId,
-                projectKey,
-                args.feature_key,
-                'manage-feature',
-            ),
     )
 }
 
@@ -548,122 +318,6 @@ export function registerFeatureTools(
             return await deleteFeatureHandler(validatedArgs, apiClient)
         },
     )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'fetch_feature_variations',
-    //     {
-    //         description: [
-    //             'Get a list of variations for a feature.',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'Get Feature Variations',
-    //             readOnlyHint: true,
-    //         },
-    //         inputSchema: ListVariationsArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = ListVariationsArgsSchema.parse(args)
-    //         return await fetchFeatureVariationsHandler(validatedArgs, apiClient)
-    //     },
-    // )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'create_feature_variation',
-    //     {
-    //         description: [
-    //             'Create a new variation within a feature.',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'Create Feature Variation',
-    //         },
-    //         inputSchema: CreateVariationArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = CreateVariationArgsSchema.parse(args)
-    //         return await createFeatureVariationHandler(validatedArgs, apiClient)
-    //     },
-    // )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'update_feature_variation',
-    //     {
-    //         description: [
-    //             'Update an existing variation by key.',
-    //             '⚠️ WARNING: Updating a feature variation may affect production environments.',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'Update Feature Variation',
-    //             destructiveHint: true,
-    //         },
-    //         inputSchema: UpdateVariationArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = UpdateVariationArgsSchema.parse(args)
-    //         return await updateFeatureVariationHandler(validatedArgs, apiClient)
-    //     },
-    // )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'set_feature_targeting',
-    //     {
-    //         description: [
-    //             'Set targeting status for a feature in an environment.',
-    //             '⚠️ IMPORTANT: Always confirm with the user before making changes to production environments (environments where type = "production").',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'Set Feature Targeting',
-    //             destructiveHint: true,
-    //         },
-    //         inputSchema: SetFeatureTargetingArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = SetFeatureTargetingArgsSchema.parse(args)
-    //         return await setFeatureTargetingHandler(validatedArgs, apiClient)
-    //     },
-    // )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'list_feature_targeting',
-    //     {
-    //         description: [
-    //             'List feature configurations (targeting rules) for a feature.',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'List Feature Targeting Rules',
-    //             readOnlyHint: true,
-    //         },
-    //         inputSchema: ListFeatureTargetingArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = ListFeatureTargetingArgsSchema.parse(args)
-    //         return await listFeatureTargetingHandler(validatedArgs, apiClient)
-    //     },
-    // )
-
-    // serverInstance.registerToolWithErrorHandling(
-    //     'update_feature_targeting',
-    //     {
-    //         description: [
-    //             'Update feature configuration (targeting rules) for a feature in an environment.',
-    //             '⚠️ IMPORTANT: Always confirm with the user before making changes to production environments (environments where type = "production").',
-    //             'Include dashboard link in the response.',
-    //         ].join('\n'),
-    //         annotations: {
-    //             title: 'Update Feature Targeting Rules',
-    //             destructiveHint: true,
-    //         },
-    //         inputSchema: UpdateFeatureTargetingArgsSchema.shape,
-    //     },
-    //     async (args: any) => {
-    //         const validatedArgs = UpdateFeatureTargetingArgsSchema.parse(args)
-    //         return await updateFeatureTargetingHandler(validatedArgs, apiClient)
-    //     },
-    // )
 
     serverInstance.registerToolWithErrorHandling(
         'get_feature_audit_log_history',
